feat: Exclude terraform files from devspace sync
Excludes the directory `monitoring/.terraform` from initial sync.

While working on copycat, I noticed this directory contained over 300MB
of data.  This is the bulkiest directory we attempt to sync when
`devspace dev`-ing that repo by a factor of 3.

I don't expect we'll be running terraform in that repo very much, so it
seems better to exclude those files from our initial sync.