chore: Promote unstable to rc (#309)
* feat: add automemlimit service activity on by default (#305)

Adds a `automemlimit` service activity that sets `GOMEMLIMIT`[1] by
default to be 90% of the available memory for the cgroup that the
current Go process is running under. This enables the garbage collector
to better be aware of how much memory the cgroup has and to take action
when getting close to it. This will prevent OOMs at the expense of
possibly more CPU cycles.

[1]: https://weaviate.io/blog/gomemlimit-a-game-changer-for-high-memory-applications

* feat: support hpa (#304)

* feat: support hpa

* fix: alignment

* fix: use local object

* fix: syntax

* fix: appropriate blocks

---------

Co-authored-by: Jared Allard <jaredallard@users.noreply.github.com>