fix(grpcClients): add ruby to root .tool-versions if enabled
Fixes `ruby` availability to be available to the entire repository if
the `ruby` grpc client is enabled. This fixes an issue where devbase
would skip configuring bundle authentication if `bundle` isn't available
in the root of the repo.

I hacked in tests by making a `testdata` folder containing a template
used to determine if we successfully included ruby or not based on the
specified arguments. This doesn't output to disk, thanks to the
`file.Skip` but is still able to be handled as a snapshot for testing.
Eventually, the new stencil testing framework will support this better.