fix: don't swallow panics
In https://github.com/getoutreach/stencil-golang/pull/121/files
a change was made to force exit the code while leaving main(),
which was also executed when a panic was being thrown, swallowing
and "hiding" it.

This change will at least print it to stderr for debugging.

Not sure whether we should instead make a formatted log entry but this
is how panics used to be surfaced before.

TESTED=not tested