fix: Add a 'team' tag to the main Datadog dashboard
Right now we don't tag the main `Terraform: {servicename}` datadog
dashboard with the team that owns the service, even though we do add a
link to that team in the description of the service.

Since we're planning to use "does the dashboard have a team" as a metric
for whether or not to clean up that dashboard, these Terraform
dashboards (which we use a lot) should have the team added.