feat(svc): create Run method to encapsulate most of whats in func main
This PR extracts most of the main function provided by stencil-golang into a
function clients can call instead.

QSS runs a number of peripheral services-- CDC, Ingest, Polaroid, SmartWorker,
etc., each of which need to set up their own main function. To do this, they
basically copy whats in stencil-golang. Wouldn't it be better if we could use a
library function instead? I think it would be, since it eases the burden when
it comes time to update any service-wide behavior.

Of course, we'll still want stencil to provide the kubernetes objects, e.g.
trace.yaml for example, and my hope is we'll stencil in depending on this
method instead of just doing what it does currently inline.